Cleaned packages, merged App.java and Main.java, corrected orchester.konzert() to orchester.auftreten(orchester) in line 27 of Main
