+++ de/hsos/prog3/ab01/ab01_2/src/dieteri/orchester/Orchester.java
@@ -29,7 +29,7 @@ public class Orchester {
 
     public URL getAudiodateiKonzert() { 
         //laut Aufgabenblatt URL?
-        return App.class.getResource(audioDateiKonzert);
+        return Main.class.getResource(audioDateiKonzert);
     }
 
     public void proben(Orchester orchester) {
