+++ de/hsos/prog3/ab01/ab01_2/src/dieteri/orchester/Probe.java
@@ -2,7 +2,7 @@ package src.dieteri.orchester;
 
 import java.io.IOException;
 
-import src.dieteri.audio.adapter.StdAudioPlayer;
+import src.dieteri.audio.StdAudioPlayer;
 import src.dieteri.audio.adapter.SimpleAudioPlayerAdapter;
 import src.dieteri.orchester.Verhalten;
 
@@ -11,13 +11,14 @@ public class Probe implements Verhalten{
     public void spielen(Orchester orchester) {
         if(orchester.getMusikerInnen() != null){
             for( MusikerIn m : orchester.getMusikerInnen()){
-                StdAudioPlayer pAdapter = SimpleAudioPlayerAdapter();
+                StdAudioPlayer pAdapter = new SimpleAudioPlayerAdapter();
                 try{
                     pAdapter.einmaligAbspielen(Main.class.getResource(m.getInstrument().getAudio()));
                 }catch(IOException e){
                 e.printStackTrace();
                 //e.getMessage();
                 System.out.println("Probe wird abgebrochen");
+                }
             }
         }
     }
